+++ script/tool/test/firebase_test_lab_command_test.dart
@@ -18,18 +18,15 @@ void main() {
   group('$FirebaseTestLabCommand', () {
     FileSystem fileSystem;
     late Directory packagesDir;
-    late List<String> printedMessages;
     late CommandRunner<void> runner;
     late RecordingProcessRunner processRunner;
 
     setUp(() {
       fileSystem = MemoryFileSystem();
       packagesDir = createPackagesDirectory(fileSystem: fileSystem);
-      printedMessages = <String>[];
       processRunner = RecordingProcessRunner();
-      final FirebaseTestLabCommand command = FirebaseTestLabCommand(packagesDir,
-          processRunner: processRunner,
-          print: (Object? message) => printedMessages.add(message.toString()));
+      final FirebaseTestLabCommand command =
+          FirebaseTestLabCommand(packagesDir, processRunner: processRunner);
 
       runner = CommandRunner<void>(
           'firebase_test_lab_command', 'Test for $FirebaseTestLabCommand');
@@ -48,32 +45,31 @@ void main() {
         'example/should_not_run_e2e.dart',
         'example/android/app/src/androidTest/MainActivityTest.java',
       ]);
-      await expectLater(
-          () => runCapturingPrint(runner, <String>['firebase-test-lab']),
-          throwsA(const TypeMatcher<ToolExit>()));
+
+      Error? commandError;
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['firebase-test-lab'], errorHandler: (Error e) {
+        commandError = e;
+      });
+
+      expect(commandError, isA<ToolExit>());
       expect(
-          printedMessages,
+          output,
           contains(
               '\nWarning: gcloud config set returned a non-zero exit code. Continuing anyway.'));
     });
 
-    test('runs e2e tests', () async {
+    test('runs integration tests', () async {
       createFakePlugin('plugin', packagesDir, extraFiles: <String>[
         'test/plugin_test.dart',
-        'test/plugin_e2e.dart',
-        'should_not_run_e2e.dart',
-        'lib/test/should_not_run_e2e.dart',
-        'example/test/plugin_e2e.dart',
-        'example/test_driver/plugin_e2e.dart',
-        'example/test_driver/plugin_e2e_test.dart',
+        'example/integration_test/bar_test.dart',
         'example/integration_test/foo_test.dart',
         'example/integration_test/should_not_run.dart',
         'example/android/gradlew',
-        'example/should_not_run_e2e.dart',
         'example/android/app/src/androidTest/MainActivityTest.java',
       ]);
 
-      await runCapturingPrint(runner, <String>[
+      final List<String> output = await runCapturingPrint(runner, <String>[
         'firebase-test-lab',
         '--device',
         'model=flame,version=29',
@@ -86,14 +82,17 @@ void main() {
       ]);
 
       expect(
-        printedMessages,
-        orderedEquals(<String>[
-          '\nRUNNING FIREBASE TEST LAB TESTS for plugin',
-          '\nFirebase project configured.',
-          '\n\n',
-          'All Firebase Test Lab tests successful!',
+        output,
+        containsAllInOrder(<Matcher>[
+          contains('Running for plugin'),
+          contains('Firebase project configured.'),
+          contains('Testing example/integration_test/bar_test.dart...'),
+          contains('Testing example/integration_test/foo_test.dart...'),
         ]),
       );
+      expect(output, isNot(contains('test/plugin_test.dart')));
+      expect(output,
+          isNot(contains('example/integration_test/should_not_run.dart')));
 
       expect(
         processRunner.recordedCalls,
@@ -111,7 +110,7 @@ void main() {
               '/packages/plugin/example/android'),
           ProcessCall(
               '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/test/plugin_e2e.dart'
+              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/integration_test/bar_test.dart'
                   .split(' '),
               '/packages/plugin/example/android'),
           ProcessCall(
@@ -121,7 +120,7 @@ void main() {
               '/packages/plugin/example'),
           ProcessCall(
               '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/test/plugin_e2e.dart'
+              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/integration_test/foo_test.dart'
                   .split(' '),
               '/packages/plugin/example/android'),
           ProcessCall(
@@ -129,16 +128,91 @@ void main() {
               'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/1/ --device model=flame,version=29 --device model=seoul,version=26'
                   .split(' '),
               '/packages/plugin/example'),
+        ]),
+      );
+    });
+
+    test('skips packages with no androidTest directory', () async {
+      createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+        'example/integration_test/foo_test.dart',
+        'example/android/gradlew',
+      ]);
+
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'firebase-test-lab',
+        '--device',
+        'model=flame,version=29',
+        '--device',
+        'model=seoul,version=26',
+        '--test-run-id',
+        'testRunId',
+        '--build-id',
+        'buildId',
+      ]);
+
+      expect(
+        output,
+        containsAllInOrder(<Matcher>[
+          contains('Running for plugin'),
+          contains('No example with androidTest directory'),
+        ]),
+      );
+      expect(output,
+          isNot(contains('Testing example/integration_test/foo_test.dart...')));
+
+      expect(
+        processRunner.recordedCalls,
+        orderedEquals(<ProcessCall>[]),
+      );
+    });
+
+    test('builds if gradlew is missing', () async {
+      createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+        'example/integration_test/foo_test.dart',
+        'example/android/app/src/androidTest/MainActivityTest.java',
+      ]);
+
+      final List<String> output = await runCapturingPrint(runner, <String>[
+        'firebase-test-lab',
+        '--device',
+        'model=flame,version=29',
+        '--device',
+        'model=seoul,version=26',
+        '--test-run-id',
+        'testRunId',
+        '--build-id',
+        'buildId',
+      ]);
+
+      expect(
+        output,
+        containsAllInOrder(<Matcher>[
+          contains('Running for plugin'),
+          contains('Running flutter build apk...'),
+          contains('Firebase project configured.'),
+          contains('Testing example/integration_test/foo_test.dart...'),
+        ]),
+      );
+
+      expect(
+        processRunner.recordedCalls,
+        orderedEquals(<ProcessCall>[
           ProcessCall(
-              '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/test_driver/plugin_e2e.dart'
-                  .split(' '),
-              '/packages/plugin/example/android'),
+            'flutter',
+            'build apk'.split(' '),
+            '/packages/plugin/example/android',
+          ),
           ProcessCall(
               'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/2/ --device model=flame,version=29 --device model=seoul,version=26'
+              'auth activate-service-account --key-file=${Platform.environment['HOME']}/gcloud-service-key.json'
                   .split(' '),
-              '/packages/plugin/example'),
+              null),
+          ProcessCall(
+              'gcloud', 'config set project flutter-infra'.split(' '), null),
+          ProcessCall(
+              '/packages/plugin/example/android/gradlew',
+              'app:assembleAndroidTest -Pverbose=true'.split(' '),
+              '/packages/plugin/example/android'),
           ProcessCall(
               '/packages/plugin/example/android/gradlew',
               'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/integration_test/foo_test.dart'
@@ -146,7 +220,7 @@ void main() {
               '/packages/plugin/example/android'),
           ProcessCall(
               'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/3/ --device model=flame,version=29 --device model=seoul,version=26'
+              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/0/ --device model=flame,version=29 --device model=seoul,version=26'
                   .split(' '),
               '/packages/plugin/example'),
         ]),
@@ -155,17 +229,8 @@ void main() {
 
     test('experimental flag', () async {
       createFakePlugin('plugin', packagesDir, extraFiles: <String>[
-        'test/plugin_test.dart',
-        'test/plugin_e2e.dart',
-        'should_not_run_e2e.dart',
-        'lib/test/should_not_run_e2e.dart',
-        'example/test/plugin_e2e.dart',
-        'example/test_driver/plugin_e2e.dart',
-        'example/test_driver/plugin_e2e_test.dart',
         'example/integration_test/foo_test.dart',
-        'example/integration_test/should_not_run.dart',
         'example/android/gradlew',
-        'example/should_not_run_e2e.dart',
         'example/android/app/src/androidTest/MainActivityTest.java',
       ]);
 
@@ -197,42 +262,12 @@ void main() {
               '/packages/plugin/example/android'),
           ProcessCall(
               '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/test/plugin_e2e.dart -Pextra-front-end-options=--enable-experiment%3Dexp1 -Pextra-gen-snapshot-options=--enable-experiment%3Dexp1'
-                  .split(' '),
-              '/packages/plugin/example/android'),
-          ProcessCall(
-              'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/0/ --device model=flame,version=29'
-                  .split(' '),
-              '/packages/plugin/example'),
-          ProcessCall(
-              '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/test/plugin_e2e.dart -Pextra-front-end-options=--enable-experiment%3Dexp1 -Pextra-gen-snapshot-options=--enable-experiment%3Dexp1'
-                  .split(' '),
-              '/packages/plugin/example/android'),
-          ProcessCall(
-              'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/1/ --device model=flame,version=29'
-                  .split(' '),
-              '/packages/plugin/example'),
-          ProcessCall(
-              '/packages/plugin/example/android/gradlew',
-              'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/test_driver/plugin_e2e.dart -Pextra-front-end-options=--enable-experiment%3Dexp1 -Pextra-gen-snapshot-options=--enable-experiment%3Dexp1'
-                  .split(' '),
-              '/packages/plugin/example/android'),
-          ProcessCall(
-              'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/2/ --device model=flame,version=29'
-                  .split(' '),
-              '/packages/plugin/example'),
-          ProcessCall(
-              '/packages/plugin/example/android/gradlew',
               'app:assembleDebug -Pverbose=true -Ptarget=/packages/plugin/example/integration_test/foo_test.dart -Pextra-front-end-options=--enable-experiment%3Dexp1 -Pextra-gen-snapshot-options=--enable-experiment%3Dexp1'
                   .split(' '),
               '/packages/plugin/example/android'),
           ProcessCall(
               'gcloud',
-              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/3/ --device model=flame,version=29'
+              'firebase test android run --type instrumentation --app build/app/outputs/apk/debug/app-debug.apk --test build/app/outputs/apk/androidTest/debug/app-debug-androidTest.apk --timeout 5m --results-bucket=gs://flutter_firebase_testlab --results-dir=plugins_android_test/plugin/buildId/testRunId/0/ --device model=flame,version=29'
                   .split(' '),
               '/packages/plugin/example'),
         ]),